refactor: changed file hierarchy for testability
