+++ src/__test__/integration/index.test.ts
@@ -1,5 +1,5 @@
 import request from 'supertest';
-import { app } from '../app';
+import { app } from '../../app';
 
 const endpointRequest = request.agent(app);
